moved GetTargetValAmounts
